Refactors Worker_Beanstalk to make it testable and adds test cases.

+++ application/libraries/Omeka/Job/Worker/Beanstalk.php
@@ -20,16 +20,10 @@ class Omeka_Job_Worker_Beanstalk
      */
     const MYSQL_TIMEOUT = 2147483;
 
-    public function __construct(Zend_Console_GetOpt $options,
+    public function __construct(Pheanstalk $pheanstalk,
                                 Omeka_Job_Factory $jobFactory,
                                 Omeka_Db $db
     ) {
-        $host = isset($options->host) ? $options->host : '127.0.0.1';
-        $pheanstalk = new Pheanstalk($host);
-        if (isset($options->queue) && $options->queue != 'default') {
-            $pheanstalk->watch($options->queue)
-                       ->ignore('default');
-        }
         $this->_pheanstalk = $pheanstalk;
         $this->_jobFactory = $jobFactory;
         $this->_db = $db;
@@ -37,14 +31,31 @@ class Omeka_Job_Worker_Beanstalk
 
     public function work()
     {
-        // Setting wait_timeout to its maximum should prevent the majority
-        // of "MySQL server has gone away" timeout errors.
-        $this->_db->query("SET SESSION wait_timeout=" . self::MYSQL_TIMEOUT);
-        $pheanJob = $this->_pheanstalk->reserve();
         try {
+            // Setting wait_timeout to its maximum should prevent the majority
+            // of "MySQL server has gone away" timeout errors.
+            $this->_db->query("SET SESSION wait_timeout=" . self::MYSQL_TIMEOUT);
+            $pheanJob = $this->_pheanstalk->reserve();
+            if (!$pheanJob) {
+                // Timeouts can occur when reserving a job, so this must be taken 
+                // into account.  No cause for alarm.
+                return;
+            }
             $omekaJob = $this->_jobFactory->from($pheanJob->getData());
+            if (!$omekaJob) {
+                throw new UnexpectedValueException(
+                    "Job factory returned null (should never happen)."
+                );
+            }
             $omekaJob->perform();
-	    $this->_pheanstalk->delete($pheanJob);
+	        $this->_pheanstalk->delete($pheanJob);
+        } catch (Zend_Db_Exception $e) {
+            // Bury any jobs with database problems aside from stale 
+            // connections, which should indicate to try the job a second time.
+            if (strpos($e->getMessage(), 'MySQL server has gone away') === false) {
+                $this->_pheanstalk->bury($pheanJob);
+            }
+            throw $e;
         } catch (Omeka_Job_Worker_InterruptException $e) {
             $this->_interrupt($omekaJob);
             throw $e;
